Смена типа с int на long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] array = {2, 6, 9, 43, 543, 54, 4, 99};
+        long[] array = {2L, 6L, 9L, 43L, 543L, 54L, 4L, 99L};
         SalesManager salesManager = new SalesManager(array);
         System.out.println(salesManager.max());
     }
